Make class part of a object's shape, fixes #10
The class needs to be part of the shape to have separate shape trees for each class. This allows us to use the shape also as guard for message sends. Otherwise, multiple classes with different sets of methods but identical sets of fields+field-types could have the same shape, and wrong methods would be activated (because the dispatch would hit on a wrong cache item).

Note, as with my original object model, each class has a seprate shape, this is not optimal for doing any kind of operation on multiple classes.

Signed-off-by: Stefan Marr <git@stefan-marr.de>